# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/messages": {
            "post": {
                "description": "returns string",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "MessageService"
                ],
                "summary": "Receive Rocket status messages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/rocket": {
            "get": {
                "description": "returns instance of rocket",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RocketService"
                ],
                "summary": "Return a rocket by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Rocket"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/rockets": {
            "get": {
                "description": "returns array of rockets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RocketService"
                ],
                "summary": "Returns all rockets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Rocket"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entities.Rocket": {
            "type": "object",
            "properties": {
                "eventCursor": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "launchDate": {
                    "type": "string"
                },
                "mission": {
                    "type": "string"
                },
                "rocketType": {
                    "type": "string"
                },
                "speed": {
                    "$ref": "#/definitions/entities.RocketSpeed"
                },
                "status": {
                    "$ref": "#/definitions/entities.RocketStatus"
                }
            }
        },
        "entities.RocketSpeed": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "number"
                },
                "max": {
                    "type": "number"
                }
            }
        },
        "entities.RocketStatus": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "reason": {
                    "type": "string"
                }
            }
        }
    }
}